Addressing Review Comments
IsBoardEmptyCheck: renamed to NumCardsOnBoardCheck, it now checks num cards on board based on given expected number
