--- dev/camoak/Assets/Scripts/Domain/Poker/Context/State/Action/Referee/GameStateCheck/IsBoardEmptyCheck.cs
@@ -1,8 +1,0 @@
-﻿namespace Camoak.Domain.Poker.Context.State.Action.Referee.GameStateCheck
-{
-    public class IsBoardEmptyCheck : IGameStateCheck
-    {
-        public bool IsSatisfied(PokerGameState gameState)
-            => gameState.BoardCards.Count == 0;
-    }
-}
