+++ dev/camoak/Assets/Scripts/Domain/Poker/Context/State/Action/Referee/GameStateCheck/NumCardsOnBoardCheck.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d1080644edbe74353ba2f2eca4e50127
+guid: 6172a5dd66a41427bb7939afed6ef155
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
